requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
aiofiles=23.2.1=pypi_0
altair=5.3.0=pypi_0
annotated-types=0.7.0=pypi_0
anyascii=0.3.2=pypi_0
anyio=4.4.0=pypi_0
appdirs=1.4.4=pypi_0
asttokens=2.4.1=pypi_0
attrs=23.2.0=pypi_0
audioread=3.0.1=pypi_0
av=10.0.0=pypi_0
babel=2.15.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
bert-extractive-summarizer=0.10.1=pypi_0
bleach=6.1.0=pypi_0
blis=0.7.11=pypi_0
boto3=1.34.143=pypi_0
botocore=1.34.143=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.7.2=h06a4308_0
cached-path=1.6.3=pypi_0
cachetools=5.3.3=pypi_0
catalogue=2.0.10=pypi_0
certifi=2024.7.4=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
cloudpathlib=0.18.1=pypi_0
cn2an=0.5.22=pypi_0
coloredlogs=15.0.1=pypi_0
confection=0.1.5=pypi_0
contourpy=1.2.1=pypi_0
cryptography=42.0.8=pypi_0
ctranslate2=3.24.0=pypi_0
cycler=0.12.1=pypi_0
cymem=2.0.8=pypi_0
cython=3.0.10=pypi_0
dateparser=1.1.8=pypi_0
decorator=4.4.2=pypi_0
deepfilterlib=0.5.6=pypi_0
deepfilternet=0.5.6=pypi_0
defusedxml=0.7.1=pypi_0
deprecated=1.2.14=pypi_0
dill=0.3.8=pypi_0
distance=0.1.3=pypi_0
dnspython=2.6.1=pypi_0
docopt=0.6.2=pypi_0
dtw-python=1.4.4=pypi_0
email-validator=2.2.0=pypi_0
eng-to-ipa=0.0.2=pypi_0
exceptiongroup=1.2.1=pypi_0
executing=2.0.1=pypi_0
fastapi=0.111.0=pypi_0
fastapi-cli=0.0.4=pypi_0
faster-whisper=0.9.0=pypi_0
fastjsonschema=2.20.0=pypi_0
ffmpeg-python=0.2.0=pypi_0
ffmpy=0.3.2=pypi_0
filelock=3.13.4=pypi_0
flatbuffers=24.3.25=pypi_0
fonttools=4.53.1=pypi_0
fsspec=2024.6.1=pypi_0
fugashi=1.3.0=pypi_0
future=1.0.0=pypi_0
g2p-en=2.1.0=pypi_0
g2pkk=0.1.2=pypi_0
google-api-core=2.19.1=pypi_0
google-auth=2.32.0=pypi_0
google-cloud-core=2.4.1=pypi_0
google-cloud-storage=2.17.0=pypi_0
google-crc32c=1.5.0=pypi_0
google-resumable-media=2.7.1=pypi_0
googleapis-common-protos=1.63.2=pypi_0
gradio=4.38.1=pypi_0
gradio-client=1.1.0=pypi_0
grpcio=1.64.1=pypi_0
gruut=2.2.3=pypi_0
gruut-ipa=0.13.0=pypi_0
gruut-lang-de=2.0.1=pypi_0
gruut-lang-en=2.0.1=pypi_0
gruut-lang-es=2.0.1=pypi_0
gruut-lang-fr=2.0.2=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.5=pypi_0
httptools=0.6.1=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.23.4=pypi_0
humanfriendly=10.0=pypi_0
idna=3.7=pypi_0
imageio=2.34.2=pypi_0
imageio-ffmpeg=0.5.1=pypi_0
importlib-resources=6.4.0=pypi_0
inflect=7.0.0=pypi_0
ipython=8.12.3=pypi_0
jaconv=0.3.4=pypi_0
jamo=0.4.1=pypi_0
jedi=0.19.1=pypi_0
jieba=0.42.1=pypi_0
jinja2=3.1.4=pypi_0
jmespath=1.0.1=pypi_0
joblib=1.4.2=pypi_0
jsonlines=1.2.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter-client=8.6.2=pypi_0
jupyter-core=5.7.2=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
kiwisolver=1.4.5=pypi_0
langcodes=3.4.0=pypi_0
langid=1.1.6=pypi_0
language-data=1.2.0=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libretranslatepy=2.1.1=pypi_0
librosa=0.9.1=pypi_0
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
llvmlite=0.43.0=pypi_0
loguru=0.7.2=pypi_0
lxml=5.2.2=pypi_0
marisa-trie=1.2.0=pypi_0
markdown=3.6=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.8.4=pypi_0
matplotlib-inline=0.1.7=pypi_0
mdurl=0.1.2=pypi_0
mecab-python3=1.0.5=pypi_0
melotts=0.1.2=pypi_0
mistune=3.0.2=pypi_0
more-itertools=10.3.0=pypi_0
moviepy=1.0.3=pypi_0
mpmath=1.3.0=pypi_0
multiprocess=0.70.16=pypi_0
murmurhash=1.0.10=pypi_0
myshell-openvoice=0.0.0=dev_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.4=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h6a678d5_0
networkx=2.8.8=pypi_0
nltk=3.8.1=pypi_0
noisereduce=3.0.2=pypi_0
num2words=0.5.12=pypi_0
numba=0.60.0=pypi_0
numpy=1.22.0=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.20.5=pypi_0
nvidia-nvjitlink-cu12=12.5.82=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
onnxruntime=1.18.1=pypi_0
openai-whisper=20231117=pypi_0
openssl=3.0.14=h5eee18b_0
orjson=3.10.6=pypi_0
packaging=23.2=pypi_0
pandas=2.0.3=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pypi_0
pathos=0.3.2=pypi_0
pexpect=4.9.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=10.4.0=pypi_0
pip=24.0=py310h06a4308_0
pipreqs=0.5.0=pypi_0
plac=1.4.3=pypi_0
platformdirs=4.2.2=pypi_0
pooch=1.8.2=pypi_0
pox=0.3.4=pypi_0
ppft=1.7.6.8=pypi_0
preshed=3.0.9=pypi_0
proces=0.1.7=pypi_0
proglog=0.1.10=pypi_0
prompt-toolkit=3.0.47=pypi_0
proto-plus=1.24.0=pypi_0
protobuf=5.27.2=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pycparser=2.22=pypi_0
pydantic=2.8.2=pypi_0
pydantic-core=2.20.1=pypi_0
pydub=0.25.1=pypi_0
pyexecjs=1.5.1=pypi_0
pygments=2.18.0=pypi_0
pykakasi=2.2.1=pypi_0
pyparsing=3.1.2=pypi_0
pypinyin=0.50.0=pypi_0
python=3.10.14=h955ad1f_1
python-crfsuite=0.9.10=pypi_0
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python-mecab-ko=1.3.7=pypi_0
python-mecab-ko-dic=2.1.1.post2=pypi_0
python-multipart=0.0.9=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=26.0.3=pypi_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
regex=2024.5.15=pypi_0
requests=2.32.3=pypi_0
resampy=0.4.3=pypi_0
rich=13.7.1=pypi_0
rpds-py=0.19.0=pypi_0
rsa=4.9=pypi_0
ruff=0.5.2=pypi_0
s3transfer=0.10.2=pypi_0
scikit-learn=1.5.1=pypi_0
scipy=1.11.4=pypi_0
semantic-version=2.10.0=pypi_0
setuptools=69.5.1=py310h06a4308_0
shellingham=1.5.4=pypi_0
six=1.16.0=pypi_0
smart-open=7.0.4=pypi_0
sniffio=1.3.1=pypi_0
soundfile=0.12.1=pypi_0
soupsieve=2.5=pypi_0
spacy=3.7.5=pypi_0
spacy-legacy=3.0.12=pypi_0
spacy-loggers=1.0.5=pypi_0
sqlite=3.45.3=h5eee18b_0
srsly=2.4.8=pypi_0
stack-data=0.6.3=pypi_0
starlette=0.37.2=pypi_0
sympy=1.13.0=pypi_0
tensorboard=2.16.2=pypi_0
tensorboard-data-server=0.7.2=pypi_0
thinc=8.2.5=pypi_0
threadpoolctl=3.5.0=pypi_0
tiktoken=0.7.0=pypi_0
tinycss2=1.3.0=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.13.3=pypi_0
tomlkit=0.12.0=pypi_0
toolz=0.12.1=pypi_0
torch=2.3.1=pypi_0
torchaudio=2.3.1+cpu=pypi_0
tornado=6.4.1=pypi_0
tqdm=4.66.4=pypi_0
traitlets=5.14.3=pypi_0
transformers=4.27.4=pypi_0
translators=5.9.2=pypi_0
triton=2.3.1=pypi_0
txtsplit=1.0.0=pypi_0
typer=0.12.3=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.1=pypi_0
tzlocal=5.2=pypi_0
ujson=5.10.0=pypi_0
unidecode=1.3.7=pypi_0
unidic=1.1.0=pypi_0
unidic-lite=1.0.8=pypi_0
urllib3=2.2.2=pypi_0
uvicorn=0.30.1=pypi_0
uvloop=0.19.0=pypi_0
wasabi=0.10.1=pypi_0
watchfiles=0.22.0=pypi_0
wavmark=0.0.3=pypi_0
wcwidth=0.2.13=pypi_0
weasel=0.4.1=pypi_0
webencodings=0.5.1=pypi_0
websockets=11.0.3=pypi_0
werkzeug=3.0.3=pypi_0
wheel=0.43.0=py310h06a4308_0
whisper-timestamped=1.14.2=pypi_0
wrapt=1.16.0=pypi_0
xz=5.4.6=h5eee18b_1
yarg=0.1.9=pypi_0
zlib=1.2.13=h5eee18b_1
